Added int printing

+++ source/externs.cpp
@@ -9,9 +9,10 @@
 #endif
 
 #include <cstdio>
+#include <cstdint>
 
 /// printd - printf that takes a double prints it as "%f\n", returning 0.
-extern "C" DLLEXPORT double writeln(double X) {
-  printf("%f\n", X);
+extern "C" DLLEXPORT int writeln(int32_t X) {
+  printf("%d\n", X);
   return 0;
 }
